refactor: add more checks in ai handler
